PLT-7709 Update plugin settings in the system console (#238)
* Update plugin settings in the system console

* Update plugin_settings.jsx

* Update plugin_management.jsx

* Update en.json

* Disable upload setting when plugins disabled

* Updating border css for plugin list
