+++ i18n/en.json
@@ -701,18 +701,24 @@
   "admin.password.requirementsDescription": "Character types required in a valid password.",
   "admin.password.symbol": "At least one symbol (e.g. \"~!@#$%^&*()\")",
   "admin.password.uppercase": "At least one uppercase letter",
-  "admin.plugin.activeTitle": "Active Plugins: ",
+  "admin.plugins.settings.title": "Configuration",
+  "admin.plugins.settings.enable": "Enable Plugins: ",
+  "admin.plugins.settings.enableDesc": "When true, enables plugins on your Mattermost server. Use plugins to integrate with third-party systems, extend functionality or customize the user interface of your Mattermost server. See <a href=\"https://about.mattermost.com/default-plugins\" target=\"_blank\">documentation</a> to learn more.",
+  "admin.plugins.settings.enableUploads": "Enable Plugin Uploads: ",
+  "admin.plugins.settings.enableUploadsDesc": "When true, enables plugin uploads by System Admins at <strong>Plugins > Management</strong>. If you do not plan to upload a plugin, set to false to control which plugins are installed on your server. See <a href=\"https://about.mattermost.com/default-plugins-uploads\" target=\"_blank\">documentation</a> to learn more.",
+  "admin.plugin.installedDesc": "Installed plugins on your Mattermost server. Pre-packaged plugins are installed by default, and can be deactivated but not removed.",
+  "admin.plugin.installedTitle": "Installed Plugins: ",
   "admin.plugin.banner": "Plugins are experimental and not recommended for use in production.",
   "admin.plugin.desc": "Description:",
   "admin.plugin.error.activate": "Unable to upload the plugin. It may conflict with another plugin on your server.",
   "admin.plugin.error.extract": "Encountered an error when extracting the plugin. Review your plugin file content and try again.",
   "admin.plugin.id": "ID:",
-  "admin.plugin.no_plugins": "No active plugins.",
+  "admin.plugin.no_plugins": "No installed plugins.",
   "admin.plugin.remove": "Remove",
   "admin.plugin.removing": "Removing...",
-  "admin.plugin.title": "Plugins (Experimental)",
+  "admin.plugin.management.title": "Management",
   "admin.plugin.upload": "Upload",
-  "admin.plugin.uploadDesc": "Upload a plugin for your Mattermost server. Adding or removing a webapp plugin requires users to refresh their browser or Desktop App before taking effect. See <a href=\"https://about.mattermost.com/default-plugins\">documentation</a> to learn more.",
+  "admin.plugin.uploadDesc": "Upload a plugin for your Mattermost server. See <a href=\"https://about.mattermost.com/default-plugin-uploads\" target=\"_blank\">documentation</a> to learn more.",
   "admin.plugin.uploadTitle": "Upload Plugin: ",
   "admin.plugin.uploading": "Uploading...",
   "admin.plugins.jira": "JIRA (Beta)",
@@ -962,7 +968,8 @@
   "admin.sidebar.oauth": "OAuth 2.0",
   "admin.sidebar.other": "OTHER",
   "admin.sidebar.password": "Password",
-  "admin.sidebar.plugins": "Plugins (Experimental)",
+  "admin.sidebar.plugins": "Plugins (Beta)",
+  "admin.sidebar.plugins.configuration": "Configuration",
   "admin.sidebar.policy": "Policy",
   "admin.sidebar.privacy": "Privacy",
   "admin.sidebar.publicLinks": "Public Links",
